requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
bilibili-api-python>=17.4.1
curl_cffi>=0.7.0
playwright>=1.45.0
websockets>=12.0
